+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @simple-data-open/logger
 
+## 0.0.2
+
+### Patch Changes
+
+- 修复 logger 类型报错
+
 ## 0.0.1
 
 ### Patch Changes