+++ tools/simple-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @simple-data-open/simple-cli
 
+## 0.0.16
+
+### Patch Changes
+
+- 修复 logger 类型报错
+- Updated dependencies
+  - @simple-data-open/utils@0.0.16
+
 ## 0.0.15
 
 ### Patch Changes